Squashed commit of the following:
commit 3ed129846d56fc11152ef181ceddb9786b622e4b
Author: Jeremy Asuncion <jeremyasuncion808@gmail.com>
Date:   Sun Jul 29 19:22:38 2018 -0700

    Added git wrapper to switch git user email and gpg program dynamically

commit e5b6585d9a0c0eb1838a54c0750b57859f11affc
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 29 18:35:50 2018 -0700

    Reverted gitconfig options

commit 85759e92a5a45472ba157044c956b55ee3805fc4
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 29 18:34:36 2018 -0700

    Added checks and conditions for macos in tmux and nvim config

commit c8e158e2213fd1c2fd0f7835da9d9adee97d1474
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 29 18:34:08 2018 -0700

    Added checks and support for macos and linux in fish scripts

commit db9e5341c574fca325112544f2d08b76f76ac66a
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 29 18:33:37 2018 -0700

    Added checks for macos in bin scripts

commit 5f0e5a94886423b382d8827790bee72f7941282a
Merge: cdcfc6a 44e1997
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 29 12:13:51 2018 -0700

    Merged master, added JAVA_HOME var

commit cdcfc6a0c20e0d0afc19a9383f466891d2189b7d
Merge: 89a78e0 15e9d88
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Thu Jul 26 14:49:05 2018 -0700

    Merge branch 'master' into sumo

commit 89a78e043ee81058f7103917109e104fb4426c0d
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Thu Jul 26 14:48:35 2018 -0700

    Added pip configurations for pip2 and pip3 on macOS

commit d182ae186e1132a7bfc692dcc6276256efa6d053
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Thu Jul 26 14:38:58 2018 -0700

    Removed type argument from desktop-mount in favor of automatic detection

commit 0e97cb52e55833259f17676f5ee76163c92cf3a5
Merge: 3a01fb2 503d6b0
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Thu Jul 26 14:38:31 2018 -0700

    Merge branch 'master' into sumo

commit 3a01fb2e3b81cb1b5017955dc62895b3ce5cd092
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 22 23:59:36 2018 -0700

    Added brew support in update script

commit 6ca8b3ef3cbab19b3755f4d6c5d23c9632ef4355
Merge: 570c11d dad070e
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 22 23:28:42 2018 -0700

    Fixed merge conflict with size and top aliases

commit 570c11d299e41ad4683a1cf1d293f171c4512516
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sun Jul 22 23:27:38 2018 -0700

    Added script to ssh into desktop based on VPN status

commit b2e1b37ff00798113b929a342d685ea2bb1d3613
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sat Jul 21 15:31:30 2018 -0700

    Removed show-notification applescript, moved applescript templates to share/vpn

commit b079391c92845392f0de8bb5f92910c446b1c98f
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sat Jul 21 15:29:49 2018 -0700

    Switched to using ssh dev-{local,remote} and removed desktop-ssh, added desktop-mount and ssh config

commit 43a686324a02240ee26c2218ca7d70e0f872bf7b
Merge: 815ddc1 76ac655
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sat Jul 21 15:12:29 2018 -0700

    Merge branch 'master' into sumo

commit 815ddc1a28e723588c153b8418ca03e67e24c911
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Sat Jul 21 15:02:49 2018 -0700

    Refactored ssh-desktop to desktop-ssh, added vpn executable

commit 29d2b4472608d84ccb4c489ff66d636969646694
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Fri Jul 20 23:02:06 2018 -0700

    Added single executable script accessing ssh to my desktop either local or remote

commit bccae7ff9eb3b25a051e1b1941833b22f2026df2
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Wed Jul 18 11:48:38 2018 -0700

    Switched readlink command for greadlink, removed keychain stuff, refactored aliases and fish_greeting, added node/yarn to PATH

commit 7adb87f7204501cc1898e475a56471e2ad4405d6
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Wed Jul 18 11:47:53 2018 -0700

    Changed nvim python hosts and tmux default-shell

commit d16a54ccfd44b69bd22aebd54d609039eb76435d
Author: Jeremy Asuncion <jasuncion@sumologic.com>
Date:   Wed Jul 18 11:47:01 2018 -0700

    Removed url rewrites and hub, changed user email

+++ config/fish/config.fish
@@ -8,10 +8,15 @@ end
 
 # Dotfiles path
 set -gx DOTFILES (
-  set config_file (status -f)
-  set config_file (readlink -e $config_file)
-  set dir (dirname $config_file)
-  readlink -e $dir/../..
+  set -l rl 'readlink'
+  if test (uname) = 'Darwin'
+    set rl 'greadlink'
+  end
+
+  set -l config_file (status -f)
+  set -l config_file (eval "$rl -e $config_file")
+  set -l dir (dirname $config_file)
+  eval "$rl -e $dir/../.."
 )
 
 # Source everything in core dir using file ordering
@@ -23,40 +28,43 @@ end
 set PATH (paths | sort -u)
 
 if status -i
-# stuff to do if a display server isn't available
-  if test -z $DISPLAY
-    # Start keychain for when the $DISPLAY variable isn't defined and fish is interactive
-    if exists keychain
-      set -l keys (
-        for f in ~/.ssh/*.pub
-          echo ~/.ssh/(basename $f .pub)
+  # Don't use keychain or automatic gpg server conf switching for macOS
+  if test (uname) != 'Darwin'
+    # stuff to do if a display server isn't available
+    if test -z $DISPLAY
+      # Start keychain for when the $DISPLAY variable isn't defined and fish is interactive
+      if exists keychain
+        set -l keys (
+          for f in ~/.ssh/*.pub
+            echo ~/.ssh/(basename $f .pub)
+          end
+        )
+
+        if test (count $keys) -gt 0
+          env SHELL=fish keychain --eval --agents ssh --quick --quiet --nogui $keys | source
         end
-      )
-
-      if test (count $keys) -gt 0
-        env SHELL=fish keychain --eval --agents ssh --quick --quiet --nogui $keys | source
       end
-    end
 
-    if test (readlink ~/.gnupg/gpg-agent.conf) = $DOTFILES/config/gnugpg/gpg-agent.conf
-      toggle-pinentry > /dev/null
-    end
-  else
-    # Have keychain kill all ssh-agent's if any are running and $DISPLAY is defined
-    if exists keychain; and test (count (keychain -l)) -gt 0
-      keychain --quiet -k all
-    end
+      if test (readlink ~/.gnupg/gpg-agent.conf) = $DOTFILES/config/gnugpg/gpg-agent.conf
+        toggle-pinentry > /dev/null
+      end
+    else
+      # Have keychain kill all ssh-agent's if any are running and $DISPLAY is defined
+      if exists keychain; and test (count (keychain -l)) -gt 0
+        keychain --quiet -k all
+      end
 
-    if test -e ~/.keychain
-      rm -rf ~/.keychain
-    end
+      if test -e ~/.keychain
+        rm -rf ~/.keychain
+      end
 
-    if test (readlink ~/.gnupg/gpg-agent.conf) = $DOTFILES/config/gnugpg/gpg-agent-tty.conf
-      toggle-pinentry > /dev/null
+      if test (readlink ~/.gnupg/gpg-agent.conf) = $DOTFILES/config/gnugpg/gpg-agent-tty.conf
+        toggle-pinentry > /dev/null
+      end
     end
   end
 
-# startup tmux or connect to existing session
+  # startup tmux or connect to existing session
   if exists tmux; and test -z $TMUX
     if tmux ls | grep -q main
       exec tmux a -t (whoami)/main
